// types/ttd.ts
export interface TTD {
  id: number;
  kode: string;
  namalap: string;
  id1: number;
  id2: number;
  id3: number;
  id4: number;
  header1: string;
  header2: string;
  header3: string;
  header4: string;
  is_id_1: number;
  is_id_2: number;
  is_id_3: number;
  is_id_4: number;
  nama1: string;
  jabatan1: string;
  nama2: string;
  jabatan2: string;
  nama3: string;
  jabatan3: string;
  nama4: string;
  jabatan4: string;
}

export interface UserParaf {
  id: number;
  nama: string;
  jabatan: string;
  nik: string;
}